Orders working. Starting, viewing and canceling orders. Worked on the front-end. Added checks for inputs for address, city, state, country and etc. to not have ; or | so the program won't error.

+++ Models/CuponModel.cs
@@ -3,7 +3,7 @@
 
         public string Name { get; set; }
         public string CuponCode { get; set; }
-        public float DiscountPercent { get; set; }
+        public decimal DiscountPercent { get; set; }
         public DateTime ExpirationDate { get; set; }
     }
 }
